refactor rpc test setup and fix clippy
